# Server/api_documentation_pl.yaml
openapi: 3.0.3
info:
  title: SNPManager - OpenAPI 3.0
  description: |-
    API aplikacji SNPManager umożliwia komunikację aplikacji klienckich z serwerem aplikacji. W poniższej dokumentacji znajduje się opis wszystkich obsługiwanych zapytań API umożliwiających bezpieczny dostęp do danych klientów.

    Repozytorium aplikacji:
    [kliknij tutaj](https://github.com/tkozl/SNPManager.git)

    ### Dostęp
    Do dostępu do większości zasobów konieczne jest załączenie do zapytania tokenu dostępowego. Token ten jest przesyłany przez serwer w odpowiedzi na dostarczenie poprawnych danych autoryzacyjnych użytkownika.

    ### Błędy zapytań
    Niektóre zapytania mogą zwrócić błąd `400` oznaczający wprowadzenie w zapytaniu nieprawidłowych danych. W takiej sytuacji w odpowiedzi serwer umieszcza id oraz opisy przyczyn odrzucenia zapytania. Informacje o możliwych błędach są umieszczone w specyfikacji poszczególnych zapytań. Lista możliwych błędów:
    - `invalid_email` - nieprawidłowy format adresu mailowego
    - `wrong_password` - błędne hasło - używane po zalogowaniu się na konto w operacjach wymagających potwierdzenia hasłem
    - `too_low_password_complexity` - zbyt niska złożoność hasła - hasło jest zbyt krótkie lub zdiagnozowano podatność na ataki słownikowe
    - `invalid_name` - podana nazwa zawiera niedozwolone znaki
    - `too_long_string` - został przekroczony maksymalny dopuszczalny rozmiar danych wejściowych
    - `already_exists` wskazany zasób już istnieje
...
  
...
servers:
...
paths:
  /account:
    post:
      tags:
        - Konta
      summary: Utworzenie nowego konta użytkownika
      description: Rejestracja w systemie nowego konta użytkownika poprzez podanie adresu mailowego, hasła oraz mechanizmu który zostanie wykorzystany do szyfrowania bazy danych (`aes-128`, `aes-192`, `aes-256`, `twofish-128`, `twofish-192`, `twofish-256`, `serpent-128`, `serpent-192`, `serpent-256`). Aby móc w pełni korzystać z konta należy potwierdzić adres email.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: securepassword
                encryptionType:
                  type: string
                  example: aes-256
        required: true
      responses:
        '201':
          description: Konto utworzone
        '400':
          description: Nieprawidłowy adres email (`invalid_email`), hasło niespełniające zasad złożoności (`too_low_password_complexity`), zbyt długi ciąg tekstowy (`too_long_string`) lub nieznany algorytm szyfrujący (`unknown_encryption_type`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
                
    put:
      tags:
        - Konta
      summary: Modyfikacja konta użytkownika
      description: Umożliwia zmianę adresu email lub hasła użytkownika. Po zmianie adresu email konieczna jest weryfikacja nowego przed dalszym korzystaniem z konta. Wszelkie zmiany wymagają podania obecnego hasła.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newMail:
                  type: string
                newPassword:
                  type: string
        required: true
      responses:
        '201':
          description: Zmiany zapisane
        '400':
          description:
            Nieprawidłowo wprowadzone dane. Możliwe ID błędów to `invalid_email`, `wrong_password`, `too_low_password_complexity`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '401':
          $ref: '#/components/responses/401Unauthorized'
      security:
      - BearerAuth: []
      
    get:
      tags:
        - Konta
      summary: Informacje o aktywnym koncie
      description: Pobranie danych koncie na które użytkownik jest aktualnie zalogowany. Zwraca adres email, stosowany algorytm szyfrowania, informację czy ustawiona jest weryfikacja dwuskładnikowa, informację czy adres email jest potwierdzony, datę założenia konta, data ostatniej zmiany hasła, ilość wpisów, listę ostatnich poprawnych logowań oraz listę ostatnich nieudanych logowań
      parameters:
        - in: query
          name: max_correct_logins
          description: Maksymalna liczba zwracanych prób zalogowania się zakończonych sukcesem. Domyślnie `10`.
          schema:
            type: integer
          required: false
        - in: query
          name: max_incorrect_logins
          description: Maksymalna liczba zwracanych nieudanych prób zalogowania się. Domyślnie `10`.
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: Pobrano dane o koncie
          content:
            application/json:
              schema:
                type: object
                properties:
                  mail:
                    type: string
                    example: user@example.com
                  encryptionType:
                    type: string
                    example: aes-256
                  is2faActive:
                    type: boolean
                  isMailActive:
                    type: boolean
                  creationDate:
                    type: integer
                    example: 1681040805
                  numberOfEntries:
                    type: integer
                  lastPasswordChange:
                    type: integer
                    example: 1681040805
                  lastAccess:
                    type: array
                    items:
                      type: object
                      properties:
                        time:
                          type: integer
                          example: 1681044405
                        ip:
                          type: string
                          example: 82.139.4.22
                  lastLoginErrors:
                    type: array
                    items:
                      type: object
                      properties:
                        time:
                          type: integer
                          example: 1681044345
                        ip:
                          type: string
                          example: 82.139.4.22
        '401':
          $ref: '#/components/responses/401Unauthorized'
      security:
        - BearerAuth: []
...
  /account/2fa:
    post:
      tags:
        - Konta
      summary: Konfiguracja mechanizmu 2fa
      description: Zapytanie umożliwia konfigurację mechanizmu 2fa. Jeśli 2fa jest już ustawione, należy najpierw wyłączyć aktywny mechanizm metodą `delete`.
      responses:
        '201':
          description: Mechanizm 2fa został aktywowany
          content:
            application/json:
              schema:
                type: object
                properties:
                  secretCode:
                    type: string
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          description: Mechanizm 2fa jest już aktywny
      security:
        - BearerAuth: []
        
    delete:
      tags:
        - Konta
      summary: Usuwanie mechanizmu 2fa
      description: Zapytanie pozwala na deaktywację obecnego mechanizmu 2fa z konta
      responses:
        '204':
          description: Mechanizm 2fa wyłączony
        '401':
          $ref: '#/components/responses/401Unauthorized'
      security:
        - BearerAuth: []
...
  /account/token:
    post:
      tags:
        - Konta
      summary: Wygenerowanie nowego tokenu dostępowego
      description: Zapytanie umożliwia przedłużenie ważności tokenu dostępowego, dopóki nie przekroczy on maksymalnego czasu ważności. Jeśli przedłużenie jest możliwe zwracany jest nowy token, a ten załączony do nagłówka zapytania jest unieważniany.
      responses:
        '201':
          description: Wygenerowano nowy token dostępowy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          description: Nie można wygenerować nowego tokenu - osiągnięto maksymalny dopuszczony czas aktywnej sesji
      security:
          - BearerAuth: []
          
    get:
      tags:
        - Konta
      summary: Aktywne tokeny dostępowe
      description: Zapytanie zwraca adresy IP mające w danej chwili dostęp do konta a także liczbę aktywnych tokenów dostępowych dla każdego z nich
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ip:
                      type: string
                      example: '83.140.2.33'
                    activeTokens:
                      type: integer
                      example: 2
        '401':
          $ref: '#/components/responses/401Unauthorized'
      security:
          - BearerAuth: []
    
    delete:
      tags:
        - Konta
      summary: Usuwanie aktywnych tokenów dostępowych
      description: Zapytanie unieważnia wszystkie aktywne tokeny dostępowe oprócz tego załączonego do nagłówka zapytania.
      responses:
        '204':
          description: Tokeny zostały unieważnione
        '401':
          $ref: '#/components/responses/401Unauthorized'
      security:
          - BearerAuth: []
  /login:
    post:
      tags:
        - Konta
      summary: Logowanie do konta użytkownika
      description: Logowanie do konta użytkownika wymaga przesłania nazwy użytkownika oraz hasła. Po poprawnym zalogowaniu zostanie odesłany token dostępowy ważny maksymalnie 12 godzin. Zostanie również odesłana informacja czy jest wymagane logowanie dwuskładnikowe. Jeśli tak, należy dosłać kod 2fa osobnym zapytaniem.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: securepassword
        required: true
      responses:
        '200':
          description: Poprawnie zalogowano
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AccessToken'
                  - type: object
                    properties:
                      is2faRequired:
                        type: boolean
        '401':
          description: Dane logowania są błędne
        '403':
          description: Odrzucone z powodu bezpieczeństwa
    delete:
      tags:
        - Konta
      summary: Wylogowanie się z konta użytkownika
      description: Wylogowanie się z konta użytkownika jest równoznaczne z unieważnieniem tokenu dostępowego załączonego do nagłówka zapytania.
      responses:
        '204':
          description: Poprawnie wylogowano
        '401':
          $ref: '#/components/responses/401Unauthorized'
      security:
        - BearerAuth: []
  /login/2fa:
    post:
      tags:
        - Konta
      summary: Weryfikacja dwuskładnikowa podczas logowania się na konto
      description: Zapytanie umożliwia dosłanie kodu 2fa w sytuacji kiedy jest on potrzebny aby uzyskać dostęp do konta. W odpowiedzi generuje nowy token dostępowy umożliwiający korzystanie z konta.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passcode:
                  type: integer
        required: true
      responses:
        '200':
          description: Poprawnie zalogowano
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '401':
          description: Podano błędny kod lub użytkownik nie jest zalogowany do systemu
        '422':
          description: Weryfikacja nie jest konieczna
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    AccessToken:
      type: object
      properties:
        token:
          type: string
        expiration:
          type: number
          example: 1694256060
    Errors:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              errorID:
                type: string
              description:
                type: string
...
  responses:
    401Unauthorized:
      description: Dostęp do zasobu wymaga autoryzacji
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer